Moved pricavy stuff to extra module
