+++ app/webFrontend/src/app/app-routing.module.ts
@@ -6,8 +6,8 @@ import {StartComponent} from './start/start.component';
 import {AboutComponent} from './about/about.component';
 import {ResetComponent} from './auth/reset/reset.component';
 import {RouterModule, Routes} from '@angular/router';
-import {LegalnoticeComponent} from './legalnotice/legalnotice.component';
-import {PrivacyComponent} from './privacy/privacy.component';
+import {LegalnoticeComponent} from './privacy/legalnotice/legalnotice.component';
+import {PrivacyComponent} from './privacy/privacynotice/privacy.component';
 import {UserSettingsComponent} from './user/user-settings/user-settings.component';
 import {AuthGuardService} from './shared/services/auth-guard.service';
 import {ActivationResendComponent} from './auth/activation-resend/activation-resend.component';
