--- app/webFrontend/src/app/legalnotice/legalnotice.component.html
@@ -1,41 +1,0 @@
-<!-- <div [innerHtml]="renderedContent"></div> -->
-
-<h1>Impressum</h1>
-
-<h3>Anbieter</h3>
-
-<span>
-  <p>Hochschule Darmstadt – Fachbereich Informatik<br>
-  Haardtring 100<br>
-  D-64295 Darmstadt</p>
-</span>
-
-<h3>Leitung:</h3>
-<p>Prof. Dr. Ute Trapp</p>
-
-
-<h3>Gestaltung und Umsetzung:</h3>
-<p>seit Master Projekt Systementwicklung SS 2017</p>
-
-<h3>Inhalt</h3>
-<span>
-  <p>Der Fachbereich Informatik ist um die Richtigkeit und Aktualität der auf dieser Internetpräsenz bereitgestellten Informationen bemüht. Der Fachbereich Informatik übernimmt jedoch keinerlei Gewähr für die Richtigkeit, Vollständigkeit und Aktualität der bereitgestellten Daten.</p>
-
-  <p>Sollten wiedererwartend doch fehlerhafte Angaben bereitgestellt werden,so übernimmt der Fachbereich Informatik keinerlei Haftung für durch die Nutzung dieser fehlerhaften Angaben enstandenen Schäden, egal ob materieller oder immaterieller Art.</p>
-</span>
-
-<h3>Links und Verweise</h3>
-<span>
-  <p>Der Fachbereich Informatik hat keinen Einfluss auf die aktuelle oder zukünftige Gestaltung der verlinkten Internetseiten. Werden nach der Linksetzung Veränderungen an den verlinkten Inhalten vorgenommen, so distanziert sich der Fachbereich Informatik ausdrücklich von diesen Inhalten und möchte sich diese auch nicht zu Eigen machen.</p>
-
-  <p>Bei den verlinkten Internetseiten, die außerhalb des Verantwortungsbereiches des Fachbereichs Informatik liegen, würde eine Haftungsverpflichtung ausschließlich in dem Fall in Kraft treten, in dem der Fachbereich Informatik von den Inhalten Kenntnis hat und es ihm technisch möglich und zumutbar wäre, die Nutzung im Falle rechtswidriger Inhalte zu verhindern.</p>
-</span>
-
-<h3>Datenschutz</h3>
-<p>Sollte auf den Internetseiten des Fachbereichs Informatik die Eingabe persönlicher Daten möglich sein, so erfolgt dies ausschliesslich auf freiwilliger Basis des Nutzers. Die dabei erfassten Daten werden ausschließlich für Verwaltungszwecke und nicht an Dritte weitergegeben.</p>
-
-<h3>Rechtswirksamkeit</h3>
-<p>Dieser Haftungsausschluss ist als Teil des Internetangebotes zu betrachten, von dem aus auf diese Seite verwiesen wurde. Sofern Teile oder einzelne Formulierungen dieses Textes der geltenden Rechtslage nicht, nicht mehr oder nicht vollständig entsprechen sollten, bleiben die übrigen Teile des Dokumentes in ihrem Inhalt und ihrer Gültigkeit davon unberührt.</p>
-
-<h3>Nutzungsbedingungen</h3>
-<p>Für die Nutzung von geli müssen Sie entweder als Student an der Hochschule Darmstadt eingeschrieben sein oder als Lehrbeauftragter im Besitz einer E-Mail Adresse sein, die auf h-da.de endet.</p>
