+++ app/webFrontend/src/app/privacy/legalnotice/legalnotice.component.html
@@ -1,0 +1,1 @@
+<div [innerHtml]="renderedContent"></div>
