+++ app/webFrontend/src/app/privacy/privacy.module.spec.ts
@@ -1,0 +1,13 @@
+import { PrivacyModule } from './privacy.module';
+
+describe('PrivacyModule', () => {
+  let privacyModule: PrivacyModule;
+
+  beforeEach(() => {
+    privacyModule = new PrivacyModule();
+  });
+
+  it('should create an instance', () => {
+    expect(privacyModule).toBeTruthy();
+  });
+});
